+++ src/main/java/com/skycat/mystical/datagen/EnglishLangProvider.java
@@ -67,6 +67,17 @@ class EnglishLangProvider extends FabricLanguageProvider {
         addConsequenceTranslation(tb, "turboMobs", "failedGetRandomEntityType", "Failed to get a random entity type, using zombie instead.");
 
         // Commands
+        addCommandText(tb, "mystical.credits", """
+                CREDITS:
+                skycatminepokie - Author
+                SuperiorTabby - Code & texture contributor
+                Phaserock - Texture contributor
+                Members of the Fabric Discord - Lots of programming help, and too many people to name. Thank you guys!
+                modmuss, player50, and the rest of the Fabric team - Fabric
+                Patbox @ Nucleoid - Server Translations API
+                Lucko - Permissions API
+                """);
+
         addCommandText(tb, "mystical.spell.delete.noSpells", "There are no active spells.");
         addCommandText(tb, "mystical.spell.new.success", "Successfully created new %s spell.");
         addCommandText(tb, "mystical.reload.success", "Successfully reloaded config and set night timer.");