+++ src/main/java/com/skycat/mystical/mixin/MinecraftServerTimer.java
@@ -11,7 +11,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.function.BooleanSupplier;
 
-// CREDIT: Daomephsta#0044 for help on fabric discord (and fabric bot too)
 @Mixin(MinecraftServer.class)
 public abstract class MinecraftServerTimer implements MinecraftServerTimerAccess {
     @Unique