+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Added /mystical spell delete all
 - Added some translations
 - illusionersReplaceEvokers checks for havens now
+- Added /mystical credits